detail 까지 Context API 적용 완료
