+++ src/pages/Router.jsx
@@ -2,20 +2,12 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Detail from "./Detail";
 import Home from "./Home";
 
-export default function Router({ memberLetterList, setMemberLetterList }) {
+export default function Router() {
   return (
     <BrowserRouter>
       <Routes>
         <Route path='/' element={<Home />} />
-        <Route
-          path='/detail/:member/:id'
-          element={
-            <Detail
-              memberLetterList={memberLetterList}
-              setMemberLetterList={setMemberLetterList}
-            />
-          }
-        />
+        <Route path='/detail/:member/:id' element={<Detail />} />
       </Routes>
     </BrowserRouter>
   );
